Refactoring

+++ src/utils.ts
@@ -25,7 +25,7 @@ export function anyToBytes(inputDatacap: string) {
   const formatDc = inputDatacap.replace(/[t]/g, "T").replace(/[b]/g, "B").replace(/[p]/g, "P").replace(/[I]/g, "i").replace(/\s*/g, "")
   const ext = formatDc.replace(/[0-9.]/g, '')
   const datacap = formatDc.replace(/[^0-9.]/g, '')
-  const bytes = byteConverter.convert(parseInt(datacap), ext, 'B')
+  const bytes = byteConverter.convert(parseFloat(datacap), ext, 'B')
   return bytes
 }
